Fix non-deterministic test InsteadOfQuestionTest.singleCorrectAnswer
The makeFake method inside the Activity class now receives
a parameter, which is a list of prohibited values for the
energy consumption, namely the consumptions of the other activities
in the question.
The InsteadOfQuestion constructor was edited to suit this change.
A couple of methods in the testing classes corresponding to
Activity and InsteadOfQuestion were changed, to allow proper
testing after the made changes.