Add a Utilities class for utility methods (duh)
This class can act as a place to contain any sort of methods we use that
are not specific behavior-wise to a particular class.  This is to keep
the code better organized and avoid potential code duplication.
Additionally, because all the methods are in a utility class, they are
made public instead of private which allows for us to test them.
Obviously you could make them public anyways, but it's not really
idomatic is it?