Refresh list of users in lobby
With this commit, every time a new user joins the lobby we update the
list of users in the lobby.  Before actually redrawing anything on the
screen we first do a check to see if there even are any new users.  Not
doing this check will cause lots of flickering.